+++ cmd/komodo/cmd/serve.go
@@ -6,15 +6,16 @@ import (
 	"github.com/spf13/cobra"
 )
 
+var (
+	default_host = "0.0.0.0"
+	default_port = "9090"
+)
+
 func init() {
 	rootCmd.AddCommand(serveCmd)
 	serveCmd.PersistentFlags().StringP("filename", "f", "", "Specify the filename of the LaTex File")
+	serveCmd.PersistentFlags().StringP("port", "p", "", "Specify the port on with komodo should start the webserver")
 }
-
-var (
-	host = "0.0.0.0"
-	port = "9090"
-)
 
 // serveCmd represents the serve command
 var serveCmd = &cobra.Command{
@@ -23,8 +24,14 @@ var serveCmd = &cobra.Command{
 	Long:  ``,
 	Run: func(cmd *cobra.Command, args []string) {
 		fname, _ := cmd.Flags().GetString("filename")
+		port, _ := cmd.Flags().GetString("port")
 
 		go watcher.WatchFile(fname, buildCommand)
-		web.StartServer(host, port, fname)
+
+		if port != "" {
+			web.StartServer(default_host, port, fname)
+		} else {
+			web.StartServer(default_host, default_port, fname)
+		}
 	},
 }