+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 * bugfix multipage textlayer rendering
 * adding welcomebanner
+* add port flag to serve command to specify the port komodo webserver should run on